tatou bon mais mauvais psnr

+++ tatouageJsteg.m
@@ -8,7 +8,7 @@ clear jsteg;
 Q = quantification(8,fq);
 Q= ones(8,8);
 %init focntion DCT pour blockproc
-fDct =@(block_struct) dct2(block_struct.data));
+fDct =@(block_struct) double(dct2(block_struct.data));
 %init focntion Quantification pour blockproc
 fQuantification=@(block_struct) round(block_struct.data./Q);
 %init fonction tatouage
